+++ apps/documentation/src/stories/misc/migration-guide/types.ts
@@ -89,7 +89,7 @@ export interface V910Checks {
     hide_automigration: boolean;
   };
   ngbootstrap: {
-    removed_components: boolean;
+    typeahead: boolean;
   };
   forms: {
     tooltip_validation: boolean;
@@ -160,6 +160,11 @@ export interface V910Checks {
     tabs_anchor_navigation: boolean;
     card_control: boolean;
     tag: boolean;
+  };
+  internet_header: {
+    update_package: boolean;
+    add_text_props: boolean;
+    remove_props: boolean;
   };
 }