+++ packages/internet-header/src/components/internal/Link.tsx
@@ -1,6 +1,7 @@
+import { state } from '@/data/store';
+import { LinkConfig } from '@/models/shared.model';
 import { FunctionalComponent, h } from '@stencil/core';
 import { JSXBase } from '@stencil/core/internal';
-import { LinkConfig } from '@/models/shared.model';
 
 export type LinkProps = {
   hiddenText?: boolean;
@@ -12,17 +13,21 @@ export const Link: FunctionalComponent<{ config: LinkConfig } & LinkProps> = ({
   hiddenText,
   ariaCurrentWhenActive,
   ...htmlAttributes
-}) => (
-  <a
-    {...htmlAttributes}
-    href={config.url}
-    aria-label={config.label}
-    aria-description={config.description}
-    aria-current={config.active ? ariaCurrentWhenActive : undefined}
-  >
-    {hiddenText ? <span class="visually-hidden">{config.text}</span> : config.text}
+}) => {
+  const isActive = state.activeLink !== null ? config === state.activeLink : config.active;
 
-    {'icon' in config && <post-icon aria-hidden="true" name={config.icon}></post-icon>}
-    {'image' in config && <img src={config.image.src} alt={config.image.alt} />}
-  </a>
-);
+  return (
+    <a
+      {...htmlAttributes}
+      href={config.url}
+      aria-label={config.label}
+      aria-description={config.description}
+      aria-current={isActive ? ariaCurrentWhenActive : undefined}
+    >
+      {hiddenText ? <span class="visually-hidden">{config.text}</span> : config.text}
+
+      {'icon' in config && <post-icon aria-hidden="true" name={config.icon}></post-icon>}
+      {'image' in config && <img src={config.image.src} alt={config.image.alt} />}
+    </a>
+  );
+};