+++ packages/internet-header/src/index.ts
@@ -4,3 +4,4 @@ export { Components, JSX } from './components';
 export { PostInternetHeader } from './components/post-internet-header/post-internet-header';
 export { PostInternetBreadcrumbs } from './components/post-internet-breadcrumbs/post-internet-breadcrumbs';
 export { PostInternetFooter } from './components/post-internet-footer/post-internet-footer';
+export { PostInternetLoginWidget } from './components/post-internet-login-widget/post-internet-login-widget';